subgraph has real nodes
