+++ src/graph/nodes/send_rescheduling_proposal_to_invitee_subgraph/analyze_message/main.py
@@ -1,0 +1,8 @@
+from src.graph.nodes.send_rescheduling_proposal_to_invitee_subgraph.analyze_message.types import AnalyzeMessageResponse
+from src.graph.nodes.send_rescheduling_proposal_to_invitee_subgraph.types import StateWithReceivedMessage
+
+
+async def analyze_message(state: StateWithReceivedMessage) -> AnalyzeMessageResponse:
+    return AnalyzeMessageResponse(
+        message_analysis="positive",
+    )
