+++ src/graph/nodes/send_rescheduling_proposal_to_invitee_subgraph/analyze_message/types.py
@@ -1,0 +1,9 @@
+from typing import Literal
+
+from pydantic import BaseModel
+
+MessageAnalysis = Literal["positive", "negative", "unknown"]
+
+
+class AnalyzeMessageResponse(BaseModel):
+    message_analysis: MessageAnalysis
