+++ src/graph/nodes/send_rescheduling_proposal_to_invitee_subgraph/receive_message/types.py
@@ -1,0 +1,6 @@
+from src.types.messaging import IncomingMessage
+from src.types.nodes import NodeResponse
+
+
+class ReceiveMessageResponse(NodeResponse):
+    received_message: IncomingMessage
