+++ src/graph/nodes/send_rescheduling_proposal_to_invitee_subgraph/send_message/types.py
@@ -1,0 +1,6 @@
+from src.types.messaging import OutgoingMessage
+from src.types.nodes import NodeResponse
+
+
+class SendMessageResponse(NodeResponse):
+    sent_message: OutgoingMessage
